added new estimators
- stability estimator: estimates “stability periods” when consumer gets
the freshest data from the network. estimation is based on calculating
mean and deviation of inter-arrival delay for consecutive frames. if
deviation is not bigger than defined threshold, it is considered stable
- rtt change estimator: estimates whether RTT value is stable (in the
same manner as stability estimator does) and tracks changes in RTT
value. RTT is considered changed value if either of these two
conditions are met:
1. either new RTT value differs from mean RTT for more than defined
percentage (20%)
2. or several consecutive small changes (above 8% but smaller than 20%)
have happened