added BufferControl class
buffer control is responsible for feeding incoming segments into the
buffer. at this point some information is retrieved from the buffer -
for instance, whether received segment is a first segment in the sample
slot. if so, this information is utilized for updating latency control
(which operates on sample level, not segment level). likewise, DRD
values are updated here (original or cached, depending on which segment
arrived - this information is also retrieved from the buffer).